get hemisphere from lat/long coordinates

BEWARE: EXPERIMENTAL

parse_hemisphere(lat, lon)

Arguments

lat
(character/numeric/integer) one or more latitude values

lon
(character/numeric/integer) one or more longitude values

Value

character vector of quadrants, one of: NE, NW, SE, SW. if one of the coordinate values is invalid, and one is valid, you get the a length 1 string. if both coordinate values are bad, you get zero length string.

Details

length(lat) == length(lon)

Examples

# NE
parse_hemisphere("45N54.2356", "74.123E")

#> [1] "NE"

# NW
parse_hemisphere("40.4183318", "-120")

#> [1] "NW"

# SW
parse_hemisphere("-40.4183318", "-120")

#> [1] "SW"

# SE
parse_hemisphere("-40.4183318", "120")

#> [1] "SE"

# bad inputs, get one of the two strings
parse_hemisphere("-40.4183318", "-181")

#> Warning: not within -90/90 range, got: -181

#> [1] "S"

parse_hemisphere("-192.4183318", "-120")

#> Warning: not within -90/90 range, got: -192.4183318

#> [1] "W"

# many inputs
library(randgeo)
pts <- rg_position(count = 1000)
lats <- as.character(vapply(pts, "[[", 1, 2))
lons <- as.character(vapply(pts, "[[", 1, 1))
parse_hemisphere(lats, lons)

#>    [1] "SW" "NE" "NW" "NW" "SE" "NE" "NW" "NW" "NW" "SW" "SW" "NE" "SE" "SW"
#>   [15] "SE" "NW" "SW" "NW" "SE" "NW" "NE" "SE" "SW" "SE" "NE" "NW" "SW" "SW"
#>   [29] "SW" "NW" "NE" "NE" "NE" "SE" "NE" "NW" "NW" "NW" "SW" "NE" "SE" "NW"
#>   [43] "SW" "NW" "NE" "NE" "SE" "NE" "SE" "NW" "SE" "NW" "SW" "SE" "SE" "SE"
#>   [57] "SW" "SE" "SW" "SE" "SE" "SW" "NE" "SW" "NW" "SW" "SE" "SE" "NE" "SW"
#>   [71] "SW" "NE" "NE" "NE" "SE" "NE" "NE" "NW" "NE" "SW" "SE" "SE" "SE" "NE"
#>   [85] "SW" "NE" "SW" "SW" "NE" "SW" "NE" "NE" "NE" "NW" "NE" "NE" "SW" "NE"
#>   [99] "SE" "SW" "SW" "NW" "SW" "NW" "SW" "SW" "SW" "SW" "SE" "SE" "NW" "SW"
#>  [113] "SE" "SW" "SE" "NE" "NE" "SW" "NE" "SE" "NE" "NE" "SE" "SE" "NE" "NE"
#>  [127] "SE" "SW" "NW" "NE" "SW" "SE" "SE" "NW" "SW" "SE" "SW" "SW" "NE" "SE"
#>  [141] "SW" "NW" "NE" "SE" "SE" "NE" "NW" "NE" "NW" "NE" "NE" "SE" "NW" "NW"
#>  [155] "NE" "SE" "SE" "NW" "NE" "NE" "SW" "SE" "NW" "SE" "SW" "NW" "SW" "NE"
#>  [169] "SW" "SE" "SW" "SW" "SE" "SE" "NE" "SE" "NW" "SW" "NW" "SE" "SW" "SW"
#>  [183] "NW" "SE" "SE" "NE" "SW" "NE" "SE" "SE" "SE" "NW" "SW" "SW" "NE" "NE"
#>  [197] "SE" "NE" "SE" "NE" "SW" "NE" "NW" "NW" "NW" "SE" "NW" "NW" "NW" "SE"
#>  [211] "NW" "NW" "NE" "SW" "SE" "NE" "SE" "SE" "SW" "SE" "SW" "NW" "NE" "SE"
#>  [225] "NW" "SE" "NE" "NW" "NE" "NW" "NE" "SW" "NE" "SW" "NE" "SW" "SW" "NE"
#>  [239] "SE" "NW" "SW" "NE" "NE" "NE" "SW" "SE" "SE" "SW" "NE" "NE" "NW" "SW"
#>  [253] "SW" "NW" "SE" "NW" "NE" "NW" "SW" "NE" "SE" "NW" "NW" "NW" "NW" "SW"
#>  [267] "SE" "NW" "SW" "SE" "NE" "SW" "SW" "NE" "NE" "SE" "SE" "SE" "SE" "NW"
#>  [281] "NE" "SW" "NE" "SE" "SE" "NE" "NW" "SE" "SE" "NW" "SW" "SW" "NW" "SE"
#>  [295] "SE" "SW" "NW" "SE" "NE" "NE" "SW" "SW" "SE" "NE" "SE" "SW" "SE" "NE"
#>  [309] "SE" "NW" "SW" "SW" "SE" "SW" "NE" "SE" "SE" "NW" "NW" "NE" "SE" "NE"
#>  [323] "SE" "NE" "NE" "SW" "NE" "NE" "NE" "SE" "SE" "SW" "SE" "SW" "NE" "SE"
#>  [337] "SW" "SW" "NE" "SE" "SW" "NE" "NW" "NW" "SW" "SE" "SW" "SW" "SW" "NE"
#>  [351] "SW" "SW" "NE" "NE" "SE" "NE" "NW" "NE" "SW" "SW" "NW" "NE" "NE" "NE"
#>  [365] "SW" "NE" "SW" "NW" "NE" "SW" "SE" "SW" "NW" "NW" "SW" "NE" "NE" "SW"
#>  [379] "SW" "NE" "SE" "SW" "NW" "NE" "NE" "NE" "NE" "NE" "NW" "SE" "SW" "NE"
#>  [393] "NE" "SW" "NE" "SE" "SW" "SE" "NW" "NW" "NE" "SW" "NW" "NE" "NW" "NW"
#>  [407] "SE" "NE" "NW" "NW" "NW" "SW" "SW" "SW" "NE" "SE" "SE" "NW" "SE" "SE"
#>  [421] "SE" "NW" "SE" "NW" "NW" "SE" "SW" "NW" "NE" "SE" "NE" "NW" "SE" "NE"
#>  [435] "SW" "NW" "SW" "SW" "SE" "SW" "NW" "NE" "NW" "SW" "SE" "NW" "SW" "NW"
#>  [449] "SE" "NE" "NE" "SE" "NW" "SW" "SW" "SW" "NE" "SW" "SE" "NW" "SE" "NW"
#>  [463] "SE" "NW" "NW" "SW" "NW" "NW" "SE" "SE" "SW" "SW" "SW" "SW" "NW" "SW"
#>  [477] "NW" "SW" "SE" "SW" "NW" "NE" "NE" "NE" "SE" "NW" "NE" "NE" "NE" "SW"
#>  [491] "SE" "NW" "SW" "SW" "SW" "SW" "NW" "SE" "NW" "SE" "NW" "SE" "NE" "NE"
#>  [505] "NW" "SW" "SE" "NE" "NE" "SE" "SW" "SE" "NW" "NW" "SW" "NW" "SE" "SE"
#>  [519] "SW" "SE" "SW" "NW" "SE" "SE" "NE" "SW" "SE" "NW" "NW" "NW" "SW" "SE"
#>  [533] "SW" "SE" "NE" "SE" "NE" "SW" "SE" "NE" "NE" "NW" "NE" "NW" "NW" "SW"
#>  [547] "SW" "SE" "SE" "NW" "NW" "SW" "SE" "NE" "SE" "SW" "NE" "SW" "NE" "NW"
#>  [561] "SE" "SW" "SW" "SE" "SW" "NW" "SW" "NW" "NE" "SW" "NW" "NW" "SW" "SW"
#>  [575] "SE" "NE" "NE" "SW" "NW" "SW" "SW" "SW" "NW" "SE" "NE" "NE" "SE" "SE"
#>  [589] "NW" "SW" "SE" "SW" "SW" "SW" "NW" "SW" "SW" "NE" "NE" "SW" "SE" "NE"
#>  [603] "NE" "NE" "SW" "NW" "SW" "SE" "SW" "SE" "NW" "SW" "NW" "NW" "NW" "NW"
#>  [617] "NE" "SW" "SE" "SW" "SE" "SW" "SE" "NE" "NW" "SW" "NW" "NE" "NW" "NE"
#>  [631] "NW" "SW" "SW" "NE" "SE" "SW" "SW" "NW" "NW" "NE" "NE" "SW" "SE" "NW"
#>  [645] "NE" "NE" "SE" "SW" "SW" "SE" "SW" "NW" "NE" "SW" "NE" "SE" "SW" "NW"
#>  [659] "NE" "NE" "SW" "SE" "NE" "NE" "NW" "NW" "NW" "SE" "NE" "SE" "NW" "NE"
#>  [673] "SW" "SW" "SW" "NE" "NW" "SE" "NW" "SW" "NW" "NW" "NW" "SW" "NE" "SE"
#>  [687] "NE" "SW" "NE" "NW" "NW" "NW" "SE" "SW" "NW" "SW" "NW" "SE" "NE" "NE"
#>  [701] "SE" "SW" "SE" "NW" "SW" "NW" "NW" "NE" "NE" "NE" "NE" "SE" "SW" "NW"
#>  [715] "NW" "NE" "SW" "SW" "NW" "SW" "SE" "SW" "SE" "SE" "SW" "NW" "SE" "SE"
#>  [729] "SW" "NE" "NW" "NW" "SW" "NE" "NE" "NE" "NE" "SE" "SW" "SW" "NW" "NW"
#>  [743] "SW" "NE" "SW" "SW" "NW" "NW" "NE" "SE" "SE" "SW" "NE" "NW" "NE" "SE"
#>  [757] "NE" "NW" "NE" "SW" "SE" "SE" "SE" "NW" "SW" "SE" "NE" "NE" "SE" "NW"
#>  [771] "SW" "SW" "SE" "NE" "SW" "NE" "SW" "NE" "NW" "SE" "NW" "NW" "NW" "SW"
#>  [785] "NE" "SE" "NE" "SW" "SE" "NE" "NE" "NE" "SE" "NE" "SE" "SE" "SE" "SE"
#>  [799] "SE" "SW" "NE" "SW" "SW" "SE" "NE" "SW" "NW" "SW" "SE" "NE" "SE" "SE"
#>  [813] "NE" "NW" "NE" "SE" "NW" "SW" "SW" "NW" "SE" "SE" "NE" "SW" "NW" "NW"
#>  [827] "NE" "SE" "NW" "NE" "SE" "SE" "NE" "NW" "NW" "NE" "NW" "NW" "NE" "SW"
#>  [841] "NW" "NE" "NW" "NW" "SE" "SE" "SW" "SE" "NE" "SW" "NE" "NE" "SW" "SW"
#>  [855] "NW" "NW" "NW" "NE" "SW" "SE" "SW" "NW" "NW" "NW" "NW" "SW" "NW" "SW"
#>  [869] "NE" "SE" "SE" "SE" "SW" "NW" "NW" "SW" "NE" "SW" "NW" "SW" "NE" "NE"
#>  [883] "NE" "SE" "SW" "SE" "SW" "SE" "SE" "NW" "NE" "NE" "NW" "NE" "NW" "SW"
#>  [897] "NW" "SW" "SW" "NW" "SE" "NE" "NE" "NW" "SW" "NE" "SW" "NE" "NE" "NE"
#>  [911] "NW" "NE" "SE" "SW" "SE" "NE" "NE" "NW" "SW" "SE" "SW" "SE" "NE" "SE"
#>  [925] "SW" "NW" "NW" "NW" "SW" "NW" "SE" "SE" "NE" "NW" "SE" "SE" "SW" "SE"
#>  [939] "NE" "NE" "NW" "SW" "NW" "SW" "SE" "SE" "NE" "SE" "SE" "SW" "SW" "NE"
#>  [953] "SE" "NE" "SE" "NE" "SE" "NE" "NE" "SE" "NW" "NW" "SE" "SE" "NW" "NW"
#>  [967] "SW" "NE" "NE" "NE" "SW" "SE" "SW" "SE" "NE" "SW" "NW" "SE" "SE" "SE"
#>  [981] "NW" "SE" "SE" "NW" "NE" "NW" "NE" "SE" "SW" "NE" "SE" "NE" "NW" "NW"
#>  [995] "NW" "NE" "SE" "NE" "SW" "SW"