extract degree, minutes, and seconds

extract degree, minutes, and seconds

pz_degree(lat = NULL, lon = NULL)

pz_minute(lat = NULL, lon = NULL)

pz_second(lat = NULL, lon = NULL)

# S3 method for pz
print(x, ...)

d(x)

m(x)

s(x)

# S3 method for pz
+(e1, e2)

# S3 method for pz
-(e1, e2)

# S3 method for pz
/(e1, e2)

# S3 method for pz
*(e1, e2)

Arguments

lat, lon
(numeric/integer/character) one or more latitude or longitude values. values are internally validated. only one of lat or lon accepted

x
(integer) an integer representing a degree, minute or second

...
print dots

e1, e2
objects of class pz, from using d(), m(), or s()

Value

pz_degree: integer, pz_minute: integer, pz_second: numeric, d: numeric, m: numeric, s: numeric (adding/subtracting these also gives numeric)

Details

Mathematics operators are exported for +, -, /, and *, but / and * are only exported with a stop message to say it's not supported; otherwise you'd be allow to divide degrees by minutes, leading to nonsense.

Examples

# extract parts of a coordinate value
pz_degree(45.23323)

#> [1] 45

pz_minute(45.23323)

#> [1] 13

pz_second(45.23323)

#> [1] 59.63012

pz_degree(lon = 178.23423)

#> [1] 178

pz_minute(lon = 178.23423)

#> [1] 14

pz_second(lon = 178.23423)

#> [1] 3.20068

pz_degree(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))

#> [1] 45 40 40

pz_minute(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))

#> [1] 13 25 25

pz_second(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))

#> [1] 59.630119  6.005895  5.992162

# invalid
pz_degree(445.23323)

#> Warning: not within -90/90 range, got: 445.23323

#> [1] NA

# add or subtract
d(31)

#> 31

m(44)

#> 44

s(3)

#> 3

d(31) + m(44)

#> 31.73333

d(31) - m(44)

#> 30.26667

d(31) + m(44) + s(59)

#> 31.74972

d(31) - m(44) + s(59)

#> 30.28306

d(-121) + m(1) + s(33)

#> -120.9742

unclass(d(31) + m(44) + s(59))

#> [1] 31.74972
#> attr(,"type")
#> [1] "deg"